client: fix renewal handling
This fixes two bugs in the renewal logic.
First, renewals can only be performed if the TGT is still valid, even if
the current time is still in the renewal window.

Second, the KDC is allowed to refuse refreshes. If the refresh fails,
the client needs to attempt to reauthenticate.

Both of these requirements come from RFC 4120 Section 2.3.